fix: common file and refactoring
