+++ common/form_utils.py
@@ -1,0 +1,14 @@
+COMMON_INPUT_CLASS = '''
+    mt-1
+    block
+    w-full
+    px-3
+    py-2
+    border
+    border-gray-300
+    rounded-md shadow-sm
+    focus:outline-none
+    focus:ring-indigo-500
+    focus:border-indigo-500
+    sm:text-sm
+'''.strip()
