+++ dashboard_app/forms.py
@@ -1,21 +1,6 @@
 from django import forms
+from common.form_utils import COMMON_INPUT_CLASS
 from .models import ActivityType, LessonDetail, Prefecture, SkiResort
-
-# 共通のCSSクラス
-COMMON_INPUT_CLASS = '''
-    mt-1
-    block
-    w-full
-    px-3
-    py-2
-    border
-    border-gray-300
-    rounded-md shadow-sm
-    focus:outline-none
-    focus:ring-indigo-500
-    focus:border-indigo-500
-    sm:text-sm
-'''
 
 def date_input_widget():
     return forms.DateInput(attrs={
